Changing interface, now as 0.2.0
