+++ qtmodern/windows/darwin.py
@@ -1,7 +1,41 @@
 from qtpy.QtWidgets import QVBoxLayout
+from qtpy.QtWidgets import QWidget, QFrame, QHBoxLayout, QLabel, QSizePolicy
+from qtpy.QtCore import Qt, Slot
 
-from qtmodern.csd.darwin import QCSDWindow
-from qtmodern.widgets import MacOSTitleBar as TitleBarWidget
+from qtmodern._csd.darwin import QCSDWindow
+
+
+class _MacOSTitleBar(QWidget):
+    """Titlebar macOS Widget."""
+
+    _HEIGHT = 22
+    """int: Height."""
+
+    def __init__(self, window, parent=None):
+        super().__init__(parent)
+
+        self.hLayout = QHBoxLayout(self)
+        self.hLayout.setContentsMargins(0, 0, 0, 0)
+
+        self.frmContent = QFrame(self)
+        self.frmContent.setObjectName('frmContent')
+        self.hLayoutContent = QHBoxLayout()
+        self.hLayoutContent.setContentsMargins(0, 0, 0, 0)
+
+        self.lblTitle = QLabel(self.frmContent)
+        self.lblTitle.setObjectName('lblTitle')
+        self.lblTitle.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Fixed)
+        self.lblTitle.setFixedHeight(self._HEIGHT)
+        self.lblTitle.setAlignment(Qt.AlignHCenter | Qt.AlignVCenter)
+        self.lblTitle.setText(window.windowTitle())
+        window.windowTitleChanged.connect(self.on_windowTitleChanged)
+        self.hLayout.addWidget(self.frmContent)
+        self.frmContent.setLayout(self.hLayoutContent)
+        self.hLayoutContent.addWidget(self.lblTitle)
+
+    @Slot(str)
+    def on_windowTitleChanged(self, title):
+        self.lblTitle.setText(title)
 
 
 class ModernWindow(QCSDWindow):
@@ -13,7 +47,7 @@ class ModernWindow(QCSDWindow):
         self.hLayout.setContentsMargins(0, 0, 0, 0)
         self.hLayout.setSpacing(0)
 
-        self.titlebar = TitleBarWidget(self, self)
+        self.titlebar = _MacOSTitleBar(self, self)
 
         self.hLayout.addWidget(self.titlebar)
         self.hLayout.addWidget(window)
